Added Event Type Code fr Table 9 ELD Mandate
7.25 Event Type from the ELD mandate dictates that Event Type Data Range: 1-7 as described on Table 9 of this appendix; Data Length: 1 character; Data Format: <Event Type> as in <C>; Disposition: Mandatory; Examples: [1], [5], [4], [7]

We may want to have EventTypeCode and EventType, depending which format the upstream telematics system sends. It should be the eventTypeCode, but in the ELD Mandate "Example of Print/Display detail log data," the Event Type/Status that is displayed is the eventType string message and not the eventTypeCode number.